More render updates

+++ Object Definitions/Sonic 1/SonLVLObjDefs/LZ/BeltActivation.cs
@@ -62,10 +62,10 @@ namespace S1ObjectDefinitions.LZ
 			if (obj.PropertyValue == 0)
 				return null;
 			
-			ushort xmin = obj.X;
-			ushort ymin = obj.Y;
-			ushort xmax = obj.X;
-			ushort ymax = obj.Y;
+			int xmin = obj.X;
+			int ymin = obj.Y;
+			int xmax = obj.X;
+			int ymax = obj.Y;
 			
 			for (int i = 1; i < obj.PropertyValue + 1; i++)
 			{
